Fix nasty bug where adding actors was getting messed up.
Apparently, there are duplicate actor names. I'm not sure if they
correspond to different actors or not. Assume not. (This is probably
wrong since there appear to be duplicates across actors and actresses.)
But I'm not sure what to do about it, other than maybe ignoring them
entirely, or coming up with a unique string our selves. But I really
don't want to go down that road.